src/requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
pandas>=2.2.0
pydantic>=2.6.0
python-dotenv>=1.0.0
